Refactored `key` feature `generate` and docs
Refactoring about naming for the `key generate` feature.

This refactoring doesn't change the user usage or interface. Added
development documentation design.

- Fixed some UT
- Added docs for developers
- Added automatically update for the documentation build for development
  guide.
