+++ docs/source/devel/repository_service_tuf.cli.key.rst
@@ -4,10 +4,10 @@ repository\_service\_tuf.cli.key package
 Submodules
 ----------
 
-repository\_service\_tuf.cli.key.generate\_key module
------------------------------------------------------
+repository\_service\_tuf.cli.key.generate module
+------------------------------------------------
 
-.. automodule:: repository_service_tuf.cli.key.generate_key
+.. automodule:: repository_service_tuf.cli.key.generate
    :members:
    :undoc-members:
    :show-inheritance:
