+++ docs/source/devel/repository_service_tuf.cli.rst
@@ -8,6 +8,7 @@ Subpackages
    :maxdepth: 4
 
    repository_service_tuf.cli.admin
+   repository_service_tuf.cli.key
 
 Module contents
 ---------------
