+++ tox.ini
@@ -45,7 +45,7 @@ allowlist_externals =
     plantuml
 commands =
     plantuml -o ../source/_static/ -tpng docs/diagrams/*.puml
-	sphinx-apidoc -o  docs/source/devel/ repository_service_tuf
+	sphinx-apidoc -f -o  docs/source/devel/ repository_service_tuf
 	sphinx-build -E -W -b html docs/source docs/build/html
 
 [gh-actions]
